+++ packages/framework/src/modules/products/products.model.ts
@@ -11,6 +11,7 @@ export class Product {
     name!: string;
     description!: string;
     shortDescription?: string;
+    variantId?: string;
     image?: Media.Media;
     price!: Price.Price;
     link!: string;