+++ packages/framework/src/modules/resources/resources.module.ts
@@ -2,6 +2,8 @@ import { HttpModule } from '@nestjs/axios';
 import { DynamicModule, Global, Module } from '@nestjs/common';
 import { Type } from '@nestjs/common/interfaces/type.interface';
 
+import { ProductService } from '../products/products.service';
+
 import { ResourceController } from './resources.controller';
 import { ResourceService } from './resources.service';
 import { ApiConfig } from '@/api-config';
@@ -11,6 +13,7 @@ import { ApiConfig } from '@/api-config';
 export class ResourceModule {
     static register(config: ApiConfig): DynamicModule {
         const service = config.integrations.resources.service;
+        const productService = config.integrations.products.service;
         const controller = config.integrations.resources.controller || ResourceController;
         const imports = config.integrations.resources.imports || [];
 
@@ -20,6 +23,10 @@ export class ResourceModule {
                 {
                     provide: ResourceService,
                     useClass: service as Type,
+                },
+                {
+                    provide: ProductService,
+                    useClass: productService as Type,
                 },
             ],
             imports: [HttpModule, ...imports],