+++ packages/integrations/medusajs/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @o2s/integrations.medusajs
 
+## 1.4.0
+
+### Minor Changes
+
+- 6949717: feat: updated resource integration - added product to the Resource model
+
+    - updated service-list and service-details blocks,
+    - added sorting and filters to mocks,
+
+### Patch Changes
+
+- Updated dependencies [6949717]
+    - @o2s/framework@1.5.0
+
 ## 1.3.0
 
 ### Minor Changes